refactor: 💡 test client wrapper function
